% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campaign_metrics}
\alias{campaign_metrics}
\title{Demo Data: Campaign Metrics}
\format{
An object of class \code{data.frame} with 449 rows and 7 columns.
}
\usage{
campaign_metrics
}
\description{
A sample of data extracted and combined from different advertising platforms:
Facebook, Google Ads, Bing, etc. Most importantly, the data show the total
amount invested in the campaigns (`ad_cost`).
}
\details{
Campaign data can be combined with \link{web_metrics} to get the full picture
of campaign performance. The most important performance metric is Return On
Advertising Spend (ROAS) calculated as revenue (from \link{web_metrics})
divided by ad cost in this data.
}
\seealso{
\link{web_metrics}, \link{email_metrics}
}
\keyword{datasets}
