% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{dm_define_dim}
\alias{dm_define_dim}
\title{Define dimension}
\usage{
dm_define_dim(dm, name = NULL, attributes = NULL)
}
\arguments{
\item{dm}{A dim model object}

\item{name}{The name of the dimension}

\item{attributes}{The columns of the dimension}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_define_dim(dm, name = "My_name")


}
}
