% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{dm_model}
\alias{dm_model}
\title{Create a dimensional model}
\usage{
dm_model(flat_table, dimension_columns, dm = NULL, dm_path = NULL)
}
\arguments{
\item{flat_table}{The data frame from which the dimensional model should be
created.}

\item{dimension_columns}{A named list with vectors of column names from
`flat_table` each of which should form a dimension table.}

\item{dm}{A `dm_model` object, i.e. an object returned by `dm_model`
or `dm_model_refresh`.}

\item{dm_path}{The file path for saving the model. If used the model will be
saved as parquet files and the the model returned will be a list of arrow
datasets.}
}
\value{
A `dm_model` object containing a list of dimension tables with primary
  keys pointing to the rows of `flat_table`.
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_model(flat_table, dimension_columns)


}
}
