% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{dm_model_refresh}
\alias{dm_model_refresh}
\title{Refresh a dimension}
\usage{
dm_model_refresh(dim_list, new_fact, dm = NULL, fact_name = NULL)
}
\arguments{
\item{dim_list}{a list of old dimensions (e.g. from dm_model_create())}

\item{new_fact}{A fact table}

\item{dm}{A model object}

\item{fact_name}{The name of the fact table}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_refresh(old_dim, new_fact)


}
}
