% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{dm_newfact}
\alias{dm_newfact}
\title{Get new facts}
\usage{
dm_newfact(dm, fct_name, flat_path, flat_name = NULL, date_col = "date")
}
\arguments{
\item{dm}{A model object}

\item{fct_name}{A fact table}

\item{flat_path}{Arguments passed to data.table::fwrite}

\item{flat_name}{A name}

\item{date_col}{The name of the date column (must be the same in both flat
and fact table)}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_save(old_dim, new_fact)


}
}
