% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{dm_refresh}
\alias{dm_refresh}
\title{Save mode as csv}
\usage{
dm_refresh(dm, new_fact_list, dm_path = NULL, ...)
}
\arguments{
\item{dm}{A model object}

\item{new_fact_list}{A fact table}

\item{dm_path}{A fact table}

\item{...}{Arguments passed to data.table::fwrite}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_save(old_dim, new_fact)


}
}
