% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{dm_refresh_dim}
\alias{dm_refresh_dim}
\title{Refresh a dimension}
\usage{
dm_refresh_dim(old_dim, new_fact)
}
\arguments{
\item{old_dim}{A dimension}

\item{new_fact}{A fact table}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_refresh(old_dim, new_fact)


}
}
