% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{dm_save}
\alias{dm_save}
\title{Refresh a dimension}
\usage{
dm_save(dm, path, partitioning = NULL)
}
\arguments{
\item{dm}{A model object}

\item{path}{A fact table}

\item{partitioning}{A fact table2}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_save(old_dim, new_fact)


}
}
