% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{dm_schema}
\alias{dm_schema}
\title{Create schema}
\usage{
dm_schema(dm, fact)
}
\arguments{
\item{dm}{A dim model object}

\item{fact}{The name of the dimension}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_schema(dm, name = "My_name")


}
}
