% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{dm_write_csv}
\alias{dm_write_csv}
\title{Save mode as csv}
\usage{
dm_write_csv(dm, path, ...)
}
\arguments{
\item{dm}{A model object}

\item{path}{A fact table}

\item{...}{Arguments passed to data.table::fwrite}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_save(old_dim, new_fact)


}
}
