% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{email_metrics}
\alias{email_metrics}
\title{Demo data: Email Send Data}
\format{
A data frame with 10,033 rows and 5 columns:
\describe{
  \item{email}{Unique identifier of the visitor.}
  \item{date}{Unique identifier of the transactions.}
  \item{sent}{The value of the transaction in USD.}
  \item{bounced}{The time of visit in minutes since 1970-01-01.}
  \item{opened}{Foreign key referring to dim_hour.}
  \item{clicked}{Foreign key referring to dim_hour.}
  }
}
\source{
Anonymized data from google analytics.
}
\usage{
email_metrics
}
\description{
A fact table showing individual visitors and their transactions on an
e-commerce website.
}
\keyword{datasets}
