% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{web_metrics}
\alias{web_metrics}
\title{Demo Data: Web Metrics}
\format{
An object of class \code{data.frame} with 1500 rows and 9 columns.
}
\usage{
web_metrics
}
\description{
A sample of data from a web analytics tool such as Google Analytics which is
typical of an online retailer. The data show how the retailer's digital
channels are performing.
}
\details{
Unique channels are identified by combining the columns `source`, `medium`
and `campaign`. Some channels are paid for by the retailer, others are not.
In the latter case the `campaign` column will show '(not set)'.

Performance is measured by the number of visits (sessions), the number of
purchases (transactions) and the total revenue a channel brings to the
website. A common performance metric is the conversion rate, which can be
calculated as the number of purchases divided by the number of sessions.

Typically a retailer would also want to calculate the ROI for each paid-for
campaign. However, web metrics data do not include information on investments
(ad costs), which is why online retailers often try to combine these data
with data pulled from the advertising platforms there are investing in:
Facebook, Google Ads, Bing, etc. (available as demo data in
 \link{campaign_metrics}).
}
\seealso{
\link{campaign_metrics}, \link{email_metrics}
}
\keyword{datasets}
