% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campaign_metrics}
\alias{campaign_metrics}
\title{Demo Data: Campaign Metrics}
\format{
An object of class \code{data.frame} with 449 rows and 7 columns.
}
\usage{
campaign_metrics
}
\description{
A sample of data extracted and combined from different advertising platforms:
Facebook, Google Ads, Bing, etc. Most importantly, the data show the total
amount of money invested in the campaigns.
}
\details{
Campaign data can be combined with [web metrics]\link{web_metrics}
The most important metric is `ad_cost` since
this is needed for ROI calculation.
}
\seealso{
\link{web_metrics}, \link{email_metrics}
}
\keyword{datasets}
