% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campaign_metrics}
\alias{campaign_metrics}
\title{Demo Data: Campaign Metrics}
\format{
An object of class \code{data.frame} with 449 rows and 9 columns.
}
\usage{
campaign_metrics
}
\description{
These data resemble data that could be extracted extracted and combined from
different advertising platforms: Facebook Ads, Google Ads, Bing Ads, etc.
Among other things, the data show the total amount of money invested in each
campaign (`ad_cost`).
}
\details{
The campaign metrics data can be combined with \link{web_metrics} to get the
full picture of campaign performance. You can merge the data using the
columns `source`, `medium` and `campaign` as key. For the purposes of this
package, these three columns will be used to create a single dimension which
is shared by two fact tables.

The most important performance metric is Return On Advertising Spend (ROAS)
calculated as `revenue` (from \link{web_metrics}) divided by `ad_cost`.
However, `impressions` and `clicks` are also worth considering.
}
\seealso{
\link{web_metrics}, \link{email_metrics}
}
\keyword{datasets}
