% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campaign_metrics}
\alias{campaign_metrics}
\title{Demo data: Campaign Metrics}
\format{
A data frame with 24 rows and 2 columns:
\describe{
  \item{date}{Primary key (unique identifier of hour).}
  \item{view_name}{Hour as a name (character type).}
  \item{country}{Hour as a name (character type).}
  \item{channel_grouping}{Hour as a name (character type).}
  \item{source}{Hour as a name (character type).}
  \item{medium}{Hour as a name (character type).}
  \item{campaign}{Hour as a name (character type).}
  \item{impressions}{Hour as a name (character type).}
  \item{ad_clicks}{Hour as a name (character type).}
  \item{ad_cost}{Hour as a name (character type).}
  \item{sessions}{Hour as a name (character type).}
  \item{transactions}{Hour as a name (character type).}
  \item{revenue}{Hour as a name (character type).}
  }
}
\source{
Anonymized data from google analytics.
}
\usage{
campaign_metrics
}
\description{
A look-up dimension related to Fact Visitors through the hour_key column.
}
\keyword{datasets}
