% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{dm_load}
\alias{dm_load}
\title{Load a dimensional model from Parquet files}
\usage{
dm_load(dm_path)
}
\arguments{
\item{dm_path}{Path to the dimensional model}
}
\value{
A `dm_model` object containing a list of dimension tables with primary
  keys pointing to the rows of `flat_table`.
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_model_load("/path/to/my/dim_model")


}
}
