% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{dm_refresh}
\alias{dm_refresh}
\title{Refresh a dimension}
\usage{
dm_refresh(dm, new_fact_list)
}
\arguments{
\item{dm}{A model object}

\item{new_fact_list}{A fact table}
}
\value{
A data frame
}
\description{
This is a description.
}
\examples{

\dontrun{

dm_refresh(old_dim, new_fact)


}
}
