% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{email_metrics}
\alias{email_metrics}
\title{Demo Data: Email Metrics}
\format{
An object of class \code{data.frame} with 376 rows and 6 columns.
}
\usage{
email_metrics
}
\description{
A fact table showing individual visitors and their transactions on an
e-commerce website.
}
\seealso{
\link{web_metrics}, \link{campaign_metrics}
}
\keyword{datasets}
