% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{web_metrics}
\alias{web_metrics}
\title{Demo Data: Web Metrics}
\format{
An object of class \code{data.frame} with 1500 rows and 9 columns.
}
\usage{
web_metrics
}
\description{
These data resemble data that could be extracted from a web analytics tool
such as Google Analytics. They show how an online retailer's digital channels
are performing.
}
\details{
To identify unique channels it is necessary to combine the columns `source`,
`medium` and `campaign`. Some channels are paid for by the retailer, others
are not. If not, the `campaign` column will show '(not set)'.

Channel performance is measured by number of visits (sessions), number of
purchases (transactions) and total revenue. A common performance metric is
the conversion rate, which can be calculated as the number of purchases
divided by the number of sessions.

Typically a retailer would also want to calculate the Return On Ad Spend
(ROAS) for each paid-for campaign. However, web metrics data do not normally
include information on spend (ad costs). As such, online retailers often try
to combine these data with data pulled from the advertising platforms there
are investing in: Facebook Ads, Google Ads, Bing Ads, etc. (available as demo
data in \link{campaign_metrics}).
}
\seealso{
\link{campaign_metrics}, \link{email_metrics}
}
\keyword{datasets}
